+++ src/PeridotPhony.php
@@ -7,6 +7,7 @@ namespace Eloquent\Phony\Peridot;
 use Eloquent\Phony\Phony;
 use Evenement\EventEmitterInterface;
 use Peridot\Core\Suite;
+use ReflectionException;
 use ReflectionFunction;
 
 /**
@@ -95,6 +96,16 @@ class PeridotPhony
     {
         $this->isScalarTypeHintSupported =
             method_exists('ReflectionParameter', 'getType');
+
+        try {
+            $function = new ReflectionFunction(function (object $a) {});
+            $parameters = $function->getParameters();
+            $this->isObjectTypeSupported = null === $parameters[0]->getClass();
+            // @codeCoverageIgnoreStart
+        } catch (ReflectionException $e) {
+            $this->isObjectTypeSupported = false;
+        }
+        // @codeCoverageIgnoreEnd
     }
 
     private function parameterArguments(array $parameters)
@@ -146,9 +157,23 @@ class PeridotPhony
                     break;
 
                 case 'array':
+                case 'iterable':
                     $argument = [];
 
                     break;
+
+            case 'object':
+                if ($this->isObjectTypeSupported) {
+                    $argument = (object) [];
+
+                    break;
+                }
+
+                // @codeCoverageIgnoreStart
+                $argument = Phony::mock('object')->get();
+
+                break;
+                // @codeCoverageIgnoreEnd
 
                 case 'stdclass':
                     $argument = (object) [];
@@ -182,4 +207,5 @@ class PeridotPhony
     }
 
     private $isScalarTypeHintSupported;
+    private $isObjectTypeSupported;
 }