+++ test/suite/PeridotPhony.spec.php
@@ -58,6 +58,7 @@ describe('PeridotPhony', function () {
                 stdClass $object,
                 callable $callable,
                 Closure $closure,
+                Generator $generator,
                 DateTime $time,
                 $bare,
                 stdClass $nullable = null
@@ -74,6 +75,7 @@ describe('PeridotPhony', function () {
             $object,
             $callable,
             $closure,
+            $generator,
             $time,
             $bare,
             $nullable
@@ -84,6 +86,7 @@ describe('PeridotPhony', function () {
         expect($callable)->to->be->an->instanceof('Eloquent\Phony\Spy\SpyVerifier');
         expect($callable())->to->be->null();
         expect($closure())->to->be->null();
+        expect(iterator_to_array($generator))->to->equal([]);
         expect($time)->to->be->an->instanceof('Eloquent\Phony\Mock\Mock');
         expect($bare)->to->be->null();
         expect($nullable)->to->be->null();
@@ -100,8 +103,7 @@ describe('PeridotPhony', function () {
                 bool $bool,
                 int $int,
                 float $float,
-                string $string,
-                Generator $generator
+                string $string
             ) {}
         );
         $this->suite->addTest($test);
@@ -114,14 +116,74 @@ describe('PeridotPhony', function () {
             $bool,
             $int,
             $float,
-            $string,
-            $generator
+            $string
         ) = $actual;
 
         expect($bool)->to->be->false();
         expect($int)->to->equal(0);
         expect($float)->to->equal(.0);
         expect($string)->to->equal('');
-        expect(iterator_to_array($generator))->to->equal([]);
     });
+
+    if (!version_compare(PHP_VERSION, '7.1.x', '<')) {
+        it('supports PHP 7.1 typehints', function () {
+            $test = new Test(
+                'test-c',
+                function (
+                    iterable $iterable
+                ) {}
+            );
+            $this->suite->addTest($test);
+            $this->subject->onSuiteStart($this->suite);
+            $actual = $test->getDefinitionArguments();
+
+            expect($actual)->to->be->an('array');
+
+            list(
+                $iterable
+            ) = $actual;
+
+            expect($iterable)->to->equal([]);
+        });
+    }
+
+    if (version_compare(PHP_VERSION, '7.2.x', '<')) {
+        it('supports the object type hint before PHP 7.2', function () {
+            $test = new Test(
+                'test-c',
+                eval('return function (object $object) {};')
+            );
+            $this->suite->addTest($test);
+            $this->subject->onSuiteStart($this->suite);
+            $actual = $test->getDefinitionArguments();
+
+            expect($actual)->to->be->an('array');
+
+            list(
+                $object,
+            ) = $actual;
+
+            expect($object)->to->be->an->instanceof('object');
+        });
+    } else {
+        it('supports PHP 7.2 typehints', function () {
+            $test = new Test(
+                'test-c',
+                function (
+                    object $object
+                ) {}
+            );
+            $this->suite->addTest($test);
+            $this->subject->onSuiteStart($this->suite);
+            $actual = $test->getDefinitionArguments();
+
+            expect($actual)->to->be->an('array');
+
+            list(
+                $object,
+            ) = $actual;
+
+            expect($object)->to->loosely->equal((object) []);
+        });
+    }
 });